bug fix in cell grouping

+++ atlas_counting/groupCellsIntoLeafRegions.m
@@ -11,7 +11,7 @@ linearIndicesL = sub2ind(size(annvol),  cellcoords(ileft, 2),  cellcoords(ileft,
 linearIndicesR = sub2ind(size(annvol), cellcoords(~ileft, 2), cellcoords(~ileft, 1), cellcoords(~ileft, 3));
 cellatlasidsR  = annvol(linearIndicesR);
 cellatlasidsL  = annvol(linearIndicesL);
-cellatlasids   = {[find(~ileft) cellatlasidsR], [find(ileft) cellatlasidsL]};
+cellatlasids   = {[find(~ileft) uint64(cellatlasidsR)], [find(ileft) uint64(cellatlasidsL)]};
 %--------------------------------------------------------------------------
 % calculate areas
 volumeareas = accumarray(reshape(annvol+1, numel(annvol), 1), 1, [Ngroups 1], @sum);
